pyproject.toml
[project]
name = "polycrossarb"
version = "0.1.0"
description = "Quantitative arbitrage system for Polymarket prediction markets"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pulp>=2.8",
    "networkx>=3.2",
    "structlog>=24.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
execution = [
    "py-clob-client>=0.1",
    "eth-account>=0.11",
    "web3>=6.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/polycrossarb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
